aider: Add command to 'Add a book from gutenberg', have it web scrape a book by title from gutenberg.org and covert to markdown document.
# Aider chat conversation:

USER: Add command to 'Add a book from gutenberg', have it web scrape a book by title from gutenberg.org and covert to markdown document.

ASSISTANT: To implement this feature, we need to add a new command to the plugin and create a new function to handle the command. We also need to create a new utility function to scrape the book from Gutenberg and convert it to markdown.

We will need to modify the following file:

- `src\main.ts`

Here are the steps:

1. Add a new command to the plugin in the `onload` method. The command will call a new method `addBookFromGutenberg`.

```typescript
src\main.ts
<<<<<<< HEAD
    this.addCommand({
      id: 'open-book-search-modal-to-insert',
      name: 'Insert the metadata',
      callback: () => this.insertMetadata(),
    });
=======
    this.addCommand({
      id: 'add-book-from-gutenberg',
      name: 'Add a book from Gutenberg',
      callback: () => this.addBookFromGutenberg(),
    });
>>>>>>> updated
```

2. Add the new method `addBookFromGutenberg` to the `BookSearchPlugin` class. This method will call a new utility function `scrapeBookFromGutenberg`.

```typescript
src\main.ts
<<<<<<< HEAD
  async saveSettings() {
    await this.saveData(this.settings);
  }
=======
  async addBookFromGutenberg(): Promise<void> {
    try {
      const bookTitle = await new Prompt({
        title: 'Enter the title of the book',
        placeholder: 'Title',
      }).open();

      const bookMarkdown = await scrapeBookFromGutenberg(bookTitle);

      const fileName = `${bookTitle.replace(/ /g, '_')}.md`;
      const filePath = `${this.settings.folder}/${fileName}`;
      await this.app.vault.create(filePath, bookMarkdown);

      new Notice(`Book "${bookTitle}" has been added.`);
    } catch (err) {
      console.warn(err);
      this.showNotice(err);
    }
  }

  async saveSettings() {
    await this.saveData(this.settings);
  }
>>>>>>> updated
```

Please note that the `scrapeBookFromGutenberg` function is not yet implemented. This function should be implemented in a new file `src\utils\gutenberg_scraper.ts`. This function should take a book title as input, scrape the book from Gutenberg, convert it to markdown, and return the markdown as a string.

Please add the `src\utils\gutenberg_scraper.ts` file to the chat so I can propose the changes to it.